# api/openapi.yaml
openapi: 3.0.0
info:
  title: Hospital Shift Management API
  description: API for managing hospital employee shifts, capabilities, and workstations
  version: 1.0.0
  contact:
    name: Support Team
servers:
  - url: http://localhost:8080/api/v1
    description: Development server
  - url: https://api.example.com/api/v1
    description: Production server
...
paths:
  /employees:
    get:
      tags:
        - Employees
      summary: List all employees
      operationId: listEmployees
      parameters:
        - name: limit
          in: query
          description: Maximum number of results
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          description: Number of results to skip
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: Paginated list of employees
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmployeePaginatedResponse'
        '500':
          $ref: '#/components/responses/InternalServerError'

    post:
      tags:
        - Employees
      summary: Create a new employee
      operationId: createEmployee
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateEmployeeRequest'
      responses:
        '201':
          description: Employee created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Employee'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          description: Email already exists
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /employees/import:
    post:
      tags:
        - Employees
      summary: Bulk-create employees from an Excel file
      description: |
        Rows follow the columns from GET /employees/template. `capabilities` and
        `available_shifts` are comma-separated lists of existing capability/shift
        names; unmatched names are reported as warnings rather than failing the row.
      operationId: importEmployees
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - file
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '200':
          description: Import summary
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ImportResult'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /employees/{employeeId}:
    get:
      tags:
        - Employees
      summary: Get an employee by ID
      operationId: getEmployeeById
      parameters:
        - name: employeeId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Employee details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmployeeDetail'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

    put:
      tags:
        - Employees
      summary: Update an employee
      operationId: updateEmployee
      parameters:
        - name: employeeId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateEmployeeRequest'
      responses:
        '200':
          description: Employee updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Employee'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Email already exists
        '500':
          $ref: '#/components/responses/InternalServerError'
  /employees/email/{email}:
    get:
      tags:
        - Employees
      summary: Get an employee by email
      operationId: getEmployeeByEmail
      parameters:
        - name: email
          in: path
          required: true
          schema:
            type: string
            format: email
      responses:
        '200':
          description: Employee details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmployeeDetail'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /employees/{employeeId}/capabilities:
    get:
      tags:
        - Employees
      summary: Get all capabilities of an employee
      operationId: getEmployeeCapabilities
      parameters:
        - name: employeeId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: List of capabilities
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Capability'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

    post:
      tags:
        - Employees
      summary: Add a capability to an employee
      operationId: addEmployeeCapability
      parameters:
        - name: employeeId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - capability_id
              properties:
                capability_id:
                  type: string
                  format: uuid
      responses:
        '201':
          description: Capability added successfully
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Employee already has this capability
        '500':
          $ref: '#/components/responses/InternalServerError'
  /employees/{employeeId}/personal-limits:
    get:
      tags:
        - Employees
      summary: Get an employee's personal limits
      description: >
        What this employee may be planned for beyond the ward's rules. An employee
        who has none saved answers with every field empty.
      operationId: getPersonalLimits
      parameters:
        - name: employeeId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: The employee's personal limits
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PersonalLimits'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags:
        - Employees
      summary: Replace an employee's personal limits
      description: >
        Replaces the limits as a whole; an empty body clears them. Weekdays are
        stored sorted and without duplicates.
      operationId: updatePersonalLimits
      parameters:
        - name: employeeId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatePersonalLimits'
      responses:
        '200':
          description: The saved limits
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PersonalLimits'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /holidays:
    get:
      tags:
        - Planner
      summary: Public holidays in a date range
      description: >-
        Holidays of the configured state. Years not stored yet are fetched from
        the configured holiday API first. A holiday runs on its shifts' Sunday
        times and counts as a weekend day.
      operationId: listHolidays
      parameters:
        - name: from
          in: query
          required: true
          schema:
            type: string
            format: date
        - name: to
          in: query
          required: true
          schema:
            type: string
            format: date
      responses:
        '200':
          description: Holidays in the range
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HolidaysResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
  /holidays/sync:
    post:
      tags:
        - Planner
      summary: Fetch holidays again from the holiday API
      operationId: syncHolidays
      parameters:
        - name: year
          in: query
          required: false
          description: Defaults to this year and the next.
          schema:
            type: integer
      responses:
        '200':
          description: The holidays now stored for those years
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HolidaysResponse'
        '503':
          description: Holidays are switched off or the holiday API is unreachable
  /employees/{employeeId}/available-shifts:
    get:
      tags:
        - Employees
      summary: Get available shifts for an employee
      operationId: getEmployeeAvailableShifts
      parameters:
        - name: employeeId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: List of available shifts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Shift'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

    post:
      tags:
        - Employees
      summary: Add an available shift to an employee
      operationId: addEmployeeAvailableShift
      parameters:
        - name: employeeId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - shift_id
              properties:
                shift_id:
                  type: string
                  format: uuid
      responses:
        '201':
          description: Shift added as available
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Shift already marked as available
        '500':
          $ref: '#/components/responses/InternalServerError'
  /shifts:
    get:
      tags:
        - Shifts
      summary: List all shifts
      operationId: listShifts
      parameters:
        - name: limit
          in: query
          description: Maximum number of results
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          description: Number of results to skip
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: List of shifts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Shift'
        '500':
          $ref: '#/components/responses/InternalServerError'

    post:
      tags:
        - Shifts
      summary: Create a new shift
      operationId: createShift
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateShiftRequest'
      responses:
        '201':
          description: Shift created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Shift'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /shifts/import:
    post:
      tags:
        - Shifts
      summary: Bulk-create shifts from an Excel file
      description: |
        Rows follow the columns from GET /shifts/template. Per-weekday start/end
        times are not part of the template and must be configured afterward.
      operationId: importShifts
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - file
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '200':
          description: Import summary
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ImportResult'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /shifts/{shiftId}:
    get:
      tags:
        - Shifts
      summary: Get a shift by ID
      operationId: getShiftById
      parameters:
        - name: shiftId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Shift details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Shift'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /shifts/{shiftId}/weekday-times:
    post:
      tags:
        - Shifts
      summary: Set start/end time for a weekday on a shift
      description: |
        Creates or updates the start_time and end_time for a specific weekday (0=Monday, 6=Sunday) on the given shift.
        If a time entry already exists for that (shiftId, weekday) pair, it will be updated.
      operationId: setShiftWeekdayTime
      parameters:
        - name: shiftId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SetWeekdayTimeRequest'
      responses:
        '200':
          description: Weekday time set successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WeekdayTime'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /shifts/{shiftId}/weekday-times/{weekday}:
    delete:
      tags:
        - Shifts
      summary: Delete a weekday time entry for a shift
      operationId: deleteShiftWeekdayTime
      parameters:
        - name: shiftId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: weekday
          in: path
          required: true
          description: Weekday number (0=Monday, 6=Sunday)
          schema:
            type: integer
            minimum: 0
            maximum: 6
      responses:
        '200':
          description: Weekday time deleted successfully
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /capabilities:
    get:
      tags:
        - Capabilities
      summary: List all capabilities
      operationId: listCapabilities
      parameters:
        - name: limit
          in: query
          description: Maximum number of results
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          description: Number of results to skip
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: List of capabilities
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Capability'
        '500':
          $ref: '#/components/responses/InternalServerError'

    post:
      tags:
        - Capabilities
      summary: Create a new capability
      operationId: createCapability
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCapabilityRequest'
      responses:
        '201':
          description: Capability created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Capability'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /capabilities/import:
    post:
      tags:
        - Capabilities
      summary: Bulk-create capabilities from an Excel file
      operationId: importCapabilities
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - file
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '200':
          description: Import summary
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ImportResult'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /capabilities/{capabilityId}:
    get:
      tags:
        - Capabilities
      summary: Get a capability by ID
      operationId: getCapabilityById
      parameters:
        - name: capabilityId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Capability details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Capability'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /workstations:
    get:
      tags:
        - Workstations
      summary: List all workstations
      operationId: listWorkstations
      parameters:
        - name: available
          in: query
          description: Filter by availability status
          schema:
            type: boolean
        - name: limit
          in: query
          description: Maximum number of results
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          description: Number of results to skip
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: List of workstations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Workstation'
        '500':
          $ref: '#/components/responses/InternalServerError'

    post:
      tags:
        - Workstations
      summary: Create a new workstation
      operationId: createWorkstation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateWorkstationRequest'
      responses:
        '201':
          description: Workstation created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Workstation'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  schemas:
...
    EmployeePaginatedResponse:
      type: object
      required:
        - data
        - total
        - limit
        - offset
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Employee'
          description: Employees for the current page
        total:
          type: integer
          description: Total number of employees
          example: 40
        limit:
          type: integer
          description: Maximum number of results returned
          example: 50
        offset:
          type: integer
          description: Number of results skipped
          example: 0
...
    CreateEmployeeRequest:
      type: object
      required:
        - name
        - email
        - monthly_working_hours
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 255
          example: "John Doe"
        email:
          type: string
          format: email
          minLength: 1
          maxLength: 255
          example: "john.doe@hospital.com"
        monthly_working_hours:
          type: number
          format: double
          description: Monthly working hours for the employee
          example: 160.0
...
  responses:
    BadRequest:
      description: Invalid request parameters or body
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'